Sétima Commit

+++ schemas/dispositivos.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel, Field
 
+
 class DispositivoCreate(BaseModel):
     unidade_consumidora_id: int
     dependencia_id: int
@@ -8,11 +9,13 @@ class DispositivoCreate(BaseModel):
     consumo: float = Field(gt=0)
     uso_diario: float = Field(ge=0, le=24)
 
+
 class DispositivoUpdate(BaseModel):
     nome: str
-    consumo: float = Field(gt=0)
-    uso_diario: float = Field(ge=0, le=24)
+    consumo: int = Field(gt=0)
+    uso_diario: int = Field(ge=0, le=24)
     tipo_id: int
+
 
 class DispositivoRead(BaseModel):
     id: int
@@ -23,10 +26,12 @@ class DispositivoRead(BaseModel):
     consumo: float
     uso_diario: float
 
+
 class DispositivoReadMany(BaseModel):
     id: int
     nome: str
     tipo_id: int
+
 
 class DispositivoReadList(BaseModel):
     dispositivos: list[DispositivoReadMany] = []
